JSON data into its own module
