+++ driver/src/price_finding/optimization_price_finder.rs
@@ -1,159 +1,16 @@
+mod data;
+
+use self::data::{Num, TokenId};
 use crate::models;
 use crate::price_finding::error::{ErrorKind, PriceFindingError};
 use crate::price_finding::price_finder_interface::{Fee, OptimizationModel, PriceFinding};
 
 use chrono::Utc;
 use log::{debug, error};
-use serde::de::Error as _;
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
-use serde_with::rust::display_fromstr;
-use std::borrow::Cow;
 use std::collections::BTreeSet;
 use std::fs::{create_dir_all, File};
 use std::io::{BufReader, BufWriter, Read, Write};
 use std::process::Command;
-
-/// An opaque token ID as understood by the contract.
-#[derive(Clone, Copy, Debug, Eq, Hash, Ord, PartialOrd, PartialEq)]
-pub struct TokenId(pub u16);
-
-impl<'de> Deserialize<'de> for TokenId {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let key = Cow::<str>::deserialize(deserializer)?;
-        if !key.starts_with('T') || key.len() != 5 {
-            return Err(D::Error::custom("Token ID must be of the form 'Txxxx'"));
-        }
-
-        let id = key[1..].parse::<u16>().map_err(D::Error::custom)?;
-        Ok(TokenId(id))
-    }
-}
-
-impl Serialize for TokenId {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        format!("T{:04}", self.0).serialize(serializer)
-    }
-}
-
-#[derive(Copy, Clone, Debug, Default, Eq, PartialEq, Deserialize, Serialize)]
-#[serde(transparent)]
-pub struct Num(#[serde(with = "display_fromstr")] pub u128);
-
-mod solver_output {
-    use super::{Num, TokenId};
-    use crate::models::Solution;
-    use serde::Deserialize;
-    use std::collections::HashMap;
-
-    #[derive(Deserialize)]
-    #[serde(rename_all = "camelCase")]
-    pub struct ExecutedOrder {
-        #[serde(default)]
-        pub exec_sell_amount: Num,
-        #[serde(default)]
-        pub exec_buy_amount: Num,
-    }
-
-    #[derive(Deserialize)]
-    #[serde(rename_all = "camelCase")]
-    pub struct Output {
-        pub orders: Vec<ExecutedOrder>,
-        pub prices: HashMap<TokenId, Option<Num>>,
-    }
-
-    impl Output {
-        pub fn to_solution(&self) -> Solution {
-            let prices = self
-                .prices
-                .iter()
-                .map(|(token, price)| (token.0, price.unwrap_or_default().0))
-                .collect();
-            let executed_sell_amounts = self
-                .orders
-                .iter()
-                .map(|order| order.exec_sell_amount.0)
-                .collect();
-            let executed_buy_amounts = self
-                .orders
-                .iter()
-                .map(|order| order.exec_buy_amount.0)
-                .collect();
-
-            Solution {
-                prices,
-                executed_sell_amounts,
-                executed_buy_amounts,
-            }
-        }
-    }
-}
-
-mod solver_input {
-    use super::{Num, TokenId};
-    use crate::models;
-    use crate::price_finding;
-    use ethcontract::H160;
-    use serde::Serialize;
-    use std::collections::{BTreeMap, BTreeSet};
-    use std::vec::Vec;
-
-    #[derive(Serialize)]
-    #[serde(rename_all = "camelCase")]
-    pub struct Fee {
-        pub token: TokenId,
-        pub ratio: f64,
-    }
-
-    impl From<&'_ price_finding::Fee> for Fee {
-        fn from(fee: &price_finding::Fee) -> Self {
-            Fee {
-                token: TokenId(fee.token),
-                ratio: fee.ratio,
-            }
-        }
-    }
-
-    #[derive(Serialize)]
-    #[serde(rename_all = "camelCase")]
-    pub struct Order {
-        #[serde(rename = "accountID")]
-        pub account_id: H160,
-        pub sell_token: TokenId,
-        pub buy_token: TokenId,
-        pub sell_amount: Num,
-        pub buy_amount: Num,
-    }
-
-    impl From<&'_ models::Order> for Order {
-        fn from(order: &models::Order) -> Self {
-            Order {
-                account_id: order.account_id,
-                sell_token: TokenId(order.sell_token),
-                buy_token: TokenId(order.buy_token),
-                sell_amount: Num(order.sell_amount),
-                buy_amount: Num(order.buy_amount),
-            }
-        }
-    }
-
-    pub type Accounts = BTreeMap<H160, BTreeMap<TokenId, Num>>;
-
-    #[derive(Serialize)]
-    #[serde(rename_all = "camelCase")]
-    pub struct Input {
-        pub tokens: BTreeSet<TokenId>,
-        pub ref_token: TokenId,
-        pub accounts: Accounts,
-        pub orders: Vec<Order>,
-        pub fee: Option<Fee>,
-    }
-}
 
 pub struct OptimisationPriceFinder {
     // default IO methods can be replaced for unit testing
@@ -185,11 +42,8 @@ fn serialize_tokens(orders: &[models::Order]) -> BTreeSet<TokenId> {
         .collect()
 }
 
-fn serialize_balances(
-    state: &models::AccountState,
-    orders: &[models::Order],
-) -> solver_input::Accounts {
-    let mut accounts = solver_input::Accounts::new();
+fn serialize_balances(state: &models::AccountState, orders: &[models::Order]) -> data::Accounts {
+    let mut accounts = data::Accounts::new();
     for order in orders {
         let token_balances = accounts.entry(order.account_id).or_default();
         for &token in &[order.buy_token, order.sell_token] {
@@ -203,7 +57,7 @@ fn serialize_balances(
 }
 
 fn deserialize_result(result: String) -> Result<models::Solution, PriceFindingError> {
-    let output: solver_output::Output = serde_json::from_str(&result)?;
+    let output: data::Output = serde_json::from_str(&result)?;
     Ok(output.to_solution())
 }
 
@@ -213,7 +67,7 @@ impl PriceFinding for OptimisationPriceFinder {
         orders: &[models::Order],
         state: &models::AccountState,
     ) -> Result<models::Solution, PriceFindingError> {
-        let input = solver_input::Input {
+        let input = data::Input {
             tokens: serialize_tokens(&orders),
             ref_token: TokenId(0),
             accounts: serialize_balances(&state, &orders),
@@ -285,27 +139,6 @@ pub mod tests {
     use std::collections::BTreeMap;
 
     #[test]
-    fn token_id_serialization() {
-        for (key, expected) in &[
-            (json!("T0000"), Some(TokenId(0))),
-            (json!("T0042"), Some(TokenId(42))),
-            (json!("T1000"), Some(TokenId(1000))),
-            (json!("T001"), None),
-            (json!("T00001"), None),
-            (json!("00001"), None),
-            (json!("Tasdf"), None),
-        ] {
-            let result = TokenId::deserialize(key);
-            if let Some(expected) = expected {
-                assert_eq!(result.unwrap(), *expected);
-                assert_eq!(json!(expected), *key);
-            } else {
-                assert!(result.is_err());
-            }
-        }
-    }
-
-    #[test]
     fn test_serialize_tokens() {
         let orders = [
             models::Order {
@@ -359,125 +192,6 @@ pub mod tests {
     }
 
     #[test]
-    fn test_failed_deserialize_result() {
-        let json = json!({
-            "The Prices": {
-                "TA": "1",
-                "TB": "2",
-            },
-        });
-        deserialize_result(json.to_string()).expect_err("Should fail to parse");
-
-        let json = json!({
-            "orders": [],
-            "prices": {
-                "tkn1": "1",
-            },
-        });
-        deserialize_result(json.to_string()).expect_err("Should fail to parse");
-
-        let json = json!({
-            "orders": [],
-            "prices": {
-                "TX": "1",
-            },
-        });
-        deserialize_result(json.to_string()).expect_err("Should fail to parse");
-
-        let json = json!({
-            "orders": [],
-            "prices": {
-                "T9999999999": "1",
-            },
-        });
-        deserialize_result(json.to_string()).expect_err("Should fail to parse");
-    }
-
-    #[test]
-    fn serialize_result_fails_if_prices_missing() {
-        let json = json!({
-            "orders": []
-        });
-        deserialize_result(json.to_string()).expect_err("Should fail to parse");
-    }
-
-    #[test]
-    fn serialize_result_fails_if_orders_missing() {
-        let json = json!({
-            "prices": {
-                "T0000": "100",
-            },
-        });
-        deserialize_result(json.to_string()).expect_err("Should fail to parse");
-    }
-
-    #[test]
-    fn serialize_result_assumes_zero_if_order_does_not_have_sell_amount() {
-        let json = json!({
-            "prices": {
-                "T0000": "100",
-            },
-            "orders": [
-                {
-                    "execBuyAmount": "0"
-                }
-            ]
-        });
-        let result = deserialize_result(json.to_string())
-            .map_err(|err| err.to_string())
-            .expect("Should not fail to parse");
-        assert_eq!(result.executed_sell_amounts[0], 0);
-    }
-
-    #[test]
-    fn serialize_result_fails_if_order_sell_volume_not_parsable() {
-        let json = json!({
-            "prices": {
-                "T0000": "100",
-            },
-            "orders": [
-                {
-                    "execSellAmount": "0a",
-                    "execBuyAmount": "0"
-                }
-            ]
-        });
-        deserialize_result(json.to_string()).expect_err("Should fail to parse");
-    }
-
-    #[test]
-    fn serialize_result_assumes_zero_if_order_does_not_have_buy_amount() {
-        let json = json!({
-            "prices": {
-                "T0000": "100",
-            },
-            "orders": [
-                {
-                    "execSellAmount": "0"
-                }
-            ]
-        });
-        let result = deserialize_result(json.to_string()).expect("Should not fail to parse");
-        assert_eq!(result.executed_buy_amounts[0], 0);
-    }
-
-    #[test]
-    fn serialize_result_fails_if_order_buy_volume_not_parsable() {
-        let json = json!({
-            "prices": {
-                "T0000": "100",
-            },
-            "orders": [
-                {
-                    "execSellAmount": "0",
-                    "execBuyAmount": "0a"
-                }
-            ]
-        });
-        deserialize_result(json.to_string()).expect_err("Should fail to parse");
-    }
-
-    #[test]
     fn test_serialize_balances() {
         let state = models::AccountState::new(
             H256::zero(),
@@ -504,7 +218,7 @@ pub mod tests {
             },
         ];
         let result = serialize_balances(&state, &orders);
-        let mut expected = solver_input::Accounts::new();
+        let mut expected = data::Accounts::new();
 
         let mut first = BTreeMap::new();
         first.insert(TokenId(1), Num(200));
@@ -545,44 +259,5 @@ pub mod tests {
         assert!(solver
             .find_prices(&orders, &AccountState::with_balance_for(&orders))
             .is_err());
-    }
-
-    #[test]
-    fn test_balance_serialization() {
-        let mut accounts = BTreeMap::new();
-
-        // Balances should end up ordered by token ID
-        let mut user1_balances = BTreeMap::new();
-        user1_balances.insert(TokenId(3), Num(100));
-        user1_balances.insert(TokenId(2), Num(100));
-        user1_balances.insert(TokenId(1), Num(100));
-        user1_balances.insert(TokenId(0), Num(100));
-
-        // Accounts should end up sorted by account ID
-        accounts.insert(
-            "4fd7c947ca0aba9d8678885e2b8c4d6a4e946984".parse().unwrap(),
-            user1_balances,
-        );
-        accounts.insert(
-            "13a0b42b9c180065510615972858bf41d1972a55".parse().unwrap(),
-            BTreeMap::new(),
-        );
-
-        let input = solver_input::Input {
-            // tokens should also end up sorted in the end
-            tokens: [TokenId(3), TokenId(2), TokenId(1), TokenId(0)]
-                .iter()
-                .copied()
-                .collect(),
-            ref_token: TokenId(0),
-            accounts,
-            orders: vec![],
-            fee: None,
-        };
-        let result = serde_json::to_string(&input).expect("Unable to serialize account state");
-        assert_eq!(
-            result,
-            r#"{"tokens":["T0000","T0001","T0002","T0003"],"refToken":"T0000","accounts":{"0x13a0b42b9c180065510615972858bf41d1972a55":{},"0x4fd7c947ca0aba9d8678885e2b8c4d6a4e946984":{"T0000":"100","T0001":"100","T0002":"100","T0003":"100"}},"orders":[],"fee":null}"#
-        );
     }
 }
